wip: move the top-level evaluator
